(memory-server) abstracted types; separated JsonMemory as impl of KnowledgeGraphMemory; accept arg for memory location
